+++ src/main/java/com/chimaera/wagubook/controller/SearchController.java
@@ -3,9 +3,9 @@ package com.chimaera.wagubook.controller;
 import com.chimaera.wagubook.entity.Member;
 import com.chimaera.wagubook.entity.Post;
 import com.chimaera.wagubook.entity.Store;
-import com.chimaera.wagubook.repository.MemberRepository;
-import com.chimaera.wagubook.repository.PostRepository;
-import com.chimaera.wagubook.repository.StoreRepository;
+import com.chimaera.wagubook.repository.member.MemberRepository;
+import com.chimaera.wagubook.repository.post.PostRepository;
+import com.chimaera.wagubook.repository.store.StoreRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
