+++ src/main/java/com/chimaera/wagubook/repository/member/MemberRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.chimaera.wagubook.repository;
+package com.chimaera.wagubook.repository.member;
 
 import com.chimaera.wagubook.entity.Member;
 import com.chimaera.wagubook.entity.QMember;
