+++ src/main/java/com/chimaera/wagubook/repository/post/PostRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.chimaera.wagubook.repository;
+package com.chimaera.wagubook.repository.post;
 
 import com.chimaera.wagubook.entity.Post;
 import com.chimaera.wagubook.entity.QPost;
