+++ src/main/java/com/chimaera/wagubook/repository/store/StoreRepository.java
@@ -1,4 +1,4 @@
-package com.chimaera.wagubook.repository;
+package com.chimaera.wagubook.repository.store;
 
 import com.chimaera.wagubook.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
