+++ src/main/java/com/chimaera/wagubook/service/MemberService.java
@@ -3,7 +3,7 @@ package com.chimaera.wagubook.service;
 import com.chimaera.wagubook.dto.LoginRequest;
 import com.chimaera.wagubook.dto.MemberRequest;
 import com.chimaera.wagubook.entity.Member;
-import com.chimaera.wagubook.repository.MemberRepository;
+import com.chimaera.wagubook.repository.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
